Add new nested table
