+++ libs/service-portal/core/src/index.ts
@@ -33,6 +33,7 @@ export * from './components/Gallery/GalleryItem'
 export * from './components/AudioPlayer/AudioPlayer'
 export * from './components/VideoPlayer/VideoPlayer'
 export * from './components/DownloadFileButtons/DownloadFileButtons'
+export * from './components/NestedTable/NestedTable'
 export * from './utils/utils'
 export * from './utils/amountFormat'
 export * from './utils/numberFormat'
