+++ libs/service-portal/health/src/components/ExpiringTable/ExpiringExpandedTableRow.tsx
@@ -1,12 +1,4 @@
-import { ExpandRow } from '@island.is/service-portal/core'
-import {
-  Box,
-  GridColumn,
-  GridContainer,
-  GridRow,
-  Text,
-} from '@island.is/island-ui/core'
-import * as styles from './ExpiringExpandedTableRow.css'
+import { ExpandRow, NestedTable } from '@island.is/service-portal/core'
 
 interface Props {
   expiring: boolean
@@ -27,26 +19,7 @@ export const ExpiringExpandedTableRow = ({
         return { value: val }
       })}
     >
-      <Box padding={2} background="blue100">
-        <GridContainer className={styles.grid}>
-          <GridRow>
-            {foldedValues.map((item, i) => (
-              <GridColumn key={i} className={styles.col} span="4/12">
-                <Box className={styles.innerCol}>
-                  <Text fontWeight="semiBold" variant="medium" as="span">
-                    {item.title}
-                  </Text>
-                </Box>
-                <Box className={styles.innerCol}>
-                  <Text variant="medium" as="span">
-                    {item.value}
-                  </Text>
-                </Box>
-              </GridColumn>
-            ))}
-          </GridRow>
-        </GridContainer>
-      </Box>
+      <NestedTable data={foldedValues} />
     </ExpandRow>
   )
 }
